+++ src/cmd_ovlpr/mod.rs
@@ -2,5 +2,4 @@
 
 pub mod covered;
 pub mod paf2ovlp;
-pub mod replace;
 pub mod restrict;