+++ src/ovlpr.rs
@@ -12,14 +12,12 @@ fn main() -> std::io::Result<()> {
         .arg_required_else_help(true)
         .subcommand(cmd_ovlpr::covered::make_subcommand())
         .subcommand(cmd_ovlpr::paf2ovlp::make_subcommand())
-        .subcommand(cmd_ovlpr::replace::make_subcommand())
         .subcommand(cmd_ovlpr::restrict::make_subcommand());
 
     // Check which subcomamnd the user ran...
     match app.get_matches().subcommand() {
         Some(("covered", sub_matches)) => cmd_ovlpr::covered::execute(sub_matches),
         Some(("paf2ovlp", sub_matches)) => cmd_ovlpr::paf2ovlp::execute(sub_matches),
-        Some(("replace", sub_matches)) => cmd_ovlpr::replace::execute(sub_matches),
         Some(("restrict", sub_matches)) => cmd_ovlpr::restrict::execute(sub_matches),
         _ => unreachable!(),
     }?;